feat: integrate new peripherals and TB flushing into emulator core
Tie together the new peripheral models and the Unicorn TB flush shim:
- Update `System` to initialize and map new peripherals
- Call `tb_flush_all()` after interrupt delivery to ensure CPU state consistency
- Improve interrupt handling logic to prevent QEMU corruption
- Update CLI and configuration loading in `main.rs` (already staged in previous step or handled)

Made-with: Cursor

+++ src/system.rs
@@ -41,7 +41,7 @@ impl<'a, 'b> System<'a, 'b> {
                 let d = self.d.clone();
                 move |uc: &mut Unicorn<'_, ()>, addr, size, value| {
                     let mut sys = System { uc: RefCell::new(uc), p: p.clone(), d: d.clone() };
-                    p.write(&mut sys, start + addr as u32, size as u8, value as u32)
+                    p.write(&mut sys, start + addr as u32, size as u8, value)
                 }
             };
